requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
mysql-connector-python==8.2.0
opencv-python==4.8.1.78
face-recognition==1.3.0
twilio==8.10.0
numpy==1.24.3
Pillow==10.0.1
bcrypt==4.0.1
python-dotenv==1.0.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
pandas==2.0.3
cryptography==41.0.7
qrcode==7.4.2
PySerial==3.5
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2
six==1.16.0
setuptools>=65.0.0